requirements.txt
streamlit==1.28.1
crewai==0.11.0
langchain==0.0.350
litellm==1.0.0
python-dotenv==1.0.0
gitpython==3.1.40
requests==2.31.0
ansible==8.6.1 